+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.28.0
+
+- **REFACTOR**: Update `ShadCalendar` `yearSelectorMinWidth` from `100` to `64` and `monthSelectorMinWidth` from `120` to `64`.
+- **REFACTOR**: Update `ShadCalendar` position of dropdown, which is now centered in the header.
+- **FIX**: Center the placeholder in `ShadAvatar`.
+- **DOCS**: Fix missing `MainAxisSize.min` in `ShadCard` notification example.
+- **DOCS**: Adjust sheet docs example, with some spacing between buttons.
+
 ## 0.27.4
 
 - **FIX**: `ShadIconButton` icon size property isn't applied (thanks to @TahaTesser).